web_app/requirements_web.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
kafka-python==2.0.2
pydantic==2.5.0
email-validator==2.1.0
jinja2==3.1.2
bcrypt==4.0.1